making recordlinker.utils a package
