+++ src/recordlinker/utils/functools.py
@@ -1,34 +1,7 @@
 import copy
 import importlib
 import inspect
-import json
-import pathlib
 import typing
-
-
-def project_root() -> pathlib.Path:
-    """
-    Returns the path to the project root directory.
-    """
-    cwd = pathlib.Path(__file__).resolve()
-    root = cwd
-
-    # FIXME: this only works when in the git project root and will fail if we install the
-    # package into the site-packages
-    while not (root / "pyproject.toml").exists():
-        if root.parent == root:
-            raise FileNotFoundError("Project root with 'pyproject.toml' not found.")
-        root = root.parent
-    return root
-
-
-def read_json(*filepaths: str) -> dict:
-    """
-    Loads a JSON file.
-    """
-    filename = pathlib.Path(project_root(), *filepaths)
-    with open(filename, "r") as fobj:
-        return json.load(fobj)
 
 
 def bind_functions(data: dict) -> dict:
@@ -127,25 +100,3 @@ def check_signature(fn: typing.Callable, expected: typing.Callable) -> bool:
 
     # Compare return type
     return _compare_types(fn_signature.return_annotation, expected_return)
-
-
-class MockTracer:
-    """
-    A no-op OTel tracer that can be used in place of a real tracer. This is useful
-    for situations where users decide to not install the otelemetry package.
-    """
-    def start_as_current_span(self, name, **kwargs):
-        """Returns a no-op span"""
-        return self
-
-    def __enter__(self):
-        """No-op for context manager entry"""
-        pass
-
-    def __exit__(self, exc_type, exc_val, exc_tb):
-        """No-op for context manager exit"""
-        pass
-
-    def start_span(self, name, **kwargs):
-        """Returns a no-op span"""
-        return self
